refactor(order_repository,models): suffix file names with "_model.dart" convention
Before moving out into separate repository packages
