+++ packages/order_repository/lib/src/models/dartgen/drug_model.dart
